# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.APIResponse:
    description: Standard API response format with metadata and context information
    properties:
      data:
        description: Response data (varies by endpoint)
      error:
        allOf:
        - $ref: '#/definitions/dto.ErrorResponse'
        description: Error details (only present when success=false)
      meta:
        allOf:
        - $ref: '#/definitions/dto.ResponseMetadata'
        description: Response metadata with request tracking
      pagination:
        allOf:
        - $ref: '#/definitions/dto.PaginationMetadata'
        description: Pagination metadata (for paginated responses)
      success:
        description: Indicates if the request was successful
        example: true
        type: boolean
    type: object
...
  dto.LoginDto:
    properties:
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    type: object
...
  model.Comment:
    properties:
      article:
        $ref: '#/definitions/model.Article'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4300
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk aplikasi blog Develapar dengan fitur lengkap untuk manajemen
    artikel, komentar, kategori, tag, bookmark, dan like. API menggunakan standard
    response format dengan metadata, request tracking, rate limiting, dan comprehensive
    error handling.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Develapar API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access token
      parameters:
      - description: Login credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success Login
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                    message:
                      type: string
                    refresh_token:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Invalid credentials
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: User login
      tags:
      - Authentication
  /auth/refresh:
    post:
      description: Refresh access token using refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: Access token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Refresh token not found
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password
      parameters:
      - description: User registration details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully registered
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    user:
                      $ref: '#/definitions/model.User'
                  type: object
              type: object
        "400":
          description: Invalid payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "409":
          description: User already exists
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Register a new user
      tags:
      - Authentication
...
  /comments/user/{user_id}:
    get:
      description: Get a list of comments by a specific user ID
      parameters:
      - description: ID of the user whose comments to retrieve
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of comments by the user
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    comments:
                      items:
                        $ref: '#/definitions/model.Comment'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid user ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get comments by user ID
      tags:
      - Comments
...
  /likes/user/{user_id}:
    get:
      description: Get a list of likes by a specific user ID
      parameters:
      - description: ID of the user whose likes to retrieve
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of likes by the user
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    likes:
                      items:
                        $ref: '#/definitions/model.Likes'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid user ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get likes by user ID
      tags:
      - Likes
...
  /users:
    get:
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    users:
                      items:
                        $ref: '#/definitions/model.User'
                      type: array
                  type: object
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get all users
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"